[Widgets] refactoring and limits table vertical size in windows (#105)
* [Widgets] refactoring and limits table vertical size in windows

* add bounding box widget

* Precomputes table size before starting the table

---------
